+++ changes/44298-fix-goquery-dependency-side-effects
@@ -1,0 +1,1 @@
+* Fixed Fleet's Docker image failing to start in Kubernetes with an `unknown userid` error, triggered by a fleetctl dependency side effect.
