+++ cmd/fleetctl/main.go
@@ -9,10 +9,17 @@ import (
 	"runtime"
 
 	"github.com/fleetdm/fleet/v4/cmd/fleetctl/fleetctl"
+	"github.com/fleetdm/fleet/v4/cmd/fleetctl/fleetctl/goquerycmd"
 	"github.com/urfave/cli/v2"
 )
 
 func main() {
+	// Register the goquery subcommand here so that the
+	// github.com/AbGuthrie/goquery/v2 dependency (and its init function) are
+	// only linked into the fleetctl binary, not into other binaries that
+	// import the fleetctl package (e.g. fleet server).
+	fleetctl.SetGoqueryRunner(goquerycmd.Run)
+
 	app := fleetctl.CreateApp(os.Stdin, os.Stdout, os.Stderr, exitErrHandler)
 	fleetctl.StashRawArgs(app, os.Args)
 	if err := app.Run(os.Args); err != nil {
